+++ oxen-rust/Cargo.toml
@@ -4,8 +4,6 @@ members = [
     "src/benches",
     "src/cli",
     "src/lib",
-    "src/oxen-test",
-    # ^^ test-only crate: never publish!
     "src/server",
 ]
 exclude = ["experiments/*"]
