--- oxen-rust/src/oxen-test/src/lib.rs
@@ -1,2 +1,0 @@
-mod s3_tests;
-mod test;
